requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
black==22.1.0
    # via -r requirements.in
certifi==2021.10.8
    # via
    #   elastic-transport
    #   requests
charset-normalizer==2.0.12
    # via requests
click==8.0.3
    # via black
elastic-transport==8.0.1
    # via elasticsearch
elasticsearch==8.0.0
    # via -r requirements.in
idna==3.3
    # via requests
isort==5.10.1
    # via -r requirements.in
mypy-extensions==0.4.3
    # via black
pathspec==0.9.0
    # via black
platformdirs==2.5.0
    # via black
python-slugify==6.0.1
    # via -r requirements.in
requests==2.27.1
    # via sodapy
sodapy==2.1.0
    # via -r requirements.in
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via black
urllib3==1.26.8
    # via
    #   elastic-transport
    #   requests
